+++ domain/src/main/java/tachiyomi/domain/library/service/LibraryPreferences.kt
@@ -193,6 +193,8 @@ class LibraryPreferences(
 
     val autoClearChapterCache: Preference<Boolean> = preferenceStore.getBoolean("auto_clear_chapter_cache", false)
 
+    val smartChapterMerging: Preference<Boolean> = preferenceStore.getBoolean("smart_chapter_merging", false)
+
     val hideMissingChapters: Preference<Boolean> = preferenceStore.getBoolean(
         "pref_hide_missing_chapter_indicators",
         false,